requirements.txt
appdirs==1.4.4
asgiref==3.3.4
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
distlib==0.3.1
Django==3.2
filelock==3.0.12
Flask==1.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-ReCaptcha==0.4.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
idna==2.10
image-slicer==2.1.1
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
numpy==1.20.1
opencv-python==4.5.2.52
pandas==1.2.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
recaptcha==1.0rc1
requests==2.25.1
scipy==1.6.3
six==1.15.0
SQLAlchemy==1.4.15
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.4
Werkzeug==1.0.1
